Updating Multiprocessing and creating Threading

+++ Threading/main.py
@@ -4,46 +4,47 @@ import datetime
 import os
 import time
 from multiprocessing import Queue
-from multiprocessing import Process
 from Queue import Empty
 import signal
 import sys
 import threading
 
 
-class Child (Process):
-    def __init__(self, chld_name, queue):
-        Process.__init__(self)
+class Child (threading.Thread):
+    def __init__(self, chld_name, queue, e):
+        threading.Thread.__init__(self)
         self.queue = queue
         self.chld_name = chld_name
-        self.event = threading.Event()
+        self.event = e
 
-    def child_signal_handler(self, signal, frame):
+    def child_signal_handler(self):
         
-        self.event.set()
-        pid = os.getpid()
+        #self.event.set()
+
+        if self.event.is_set():
+            pid = os.getpid()
         
-        isClosed = os.kill(pid, 0)
+            isClosed = os.kill(pid, 0)
 
-        if isClosed == None:
+            if isClosed == None:
 
-            print "Child %s with pid: %d closed!" % (self.chld_name, pid)
-            self.event.clear()
-            sys.exit(0)
+                print "Child %s with pid: %d closed!" % (self.chld_name, pid)
+                self.event.clear()
+                sys.exit(0)
         
-        else:
+            else:
 
-            print "Child still working..."    
+                print "Child still working..."    
     
         
 
     def run(self):
 
-        signal.signal(signal.SIGTERM, self.child_signal_handler)
-
+       # 
         print "Starting " + self.name
         
         while not self.event.is_set():
+           # self.child_signal_handler()
             #while self.queue:
             try:
                 q = self.queue.get()
@@ -59,10 +60,17 @@ class Child (Process):
 class Parent(object):
     def __init__(self):
         self.q = Queue()
-        self.child_1 = Child("Child-1", self.q)
-        self.child_2 = Child("Child-2", self.q)
+        
+        self.event = threading.Event()
+
+        self.child_1 = Child("Child-1", self.q, self.event)
+        self.child_2 = Child("Child-2", self.q, self.event)
+        
         signal.signal(signal.SIGTERM, self.signal_term_handler)
+        
+        
         self.cur_pid = os.getpid()
+
 
     def writeToQueue(self):
         
@@ -80,10 +88,10 @@ class Parent(object):
 
             #for child in self.childlist:
             print "Terminate child: %s" % (self.child_1.name)
-            self.child_1.terminate()
+            #self.child_1.quit()
 
             print "Terminate child: %s" % (self.child_2.name)
-            self.child_2.terminate()
+            #self.child_2.quit()
             sys.exit(0)
             #os.kill(l)
 
@@ -91,7 +99,8 @@ class Parent(object):
             pass
 
     def signal_term_handler(self, signal, frame):
-
+        
+        self.event.set()
         self.termProc = True
         print 'got SIGTERM, Event started'
         self.TerminateAll()
